Change admin user management file structure
